requirements.txt
annotated-types==0.7.0
anyio==4.8.0
api==0.0.7
asgiref==3.7.2
asttokens==3.0.0
asyncpg==0.30.0
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
comm==0.2.2
debugpy==1.8.12
decorator==5.1.1
Django==5.0.1
django-cors-headers==4.3.1
django-filter==24.3
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
dnspython==2.7.0
executing==2.2.0
fastapi==0.115.7
fpdf==1.7.2
greenlet==3.1.1
h11==0.14.0
httptools==0.6.4
idna==3.10
ipykernel==6.29.5
ipython==8.31.0
itsdangerous==2.2.0
jedi==0.19.2
jupyter_client==8.6.3
jupyter_core==5.7.2
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
mdurl==0.1.2
nest-asyncio==1.6.0
nose==1.3.7
orjson==3.10.15
packaging==24.2
parso==0.8.4
passlib==1.7.4
pillow==11.1.0
platformdirs==4.3.6
prompt_toolkit==3.0.50
psutil==6.1.1
psycopg2-binary==2.9.9
pure_eval==0.2.3
pydantic==2.10.6
pydantic_core==2.27.2
Pygments==2.19.1
PyJWT==2.8.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.20
pytz==2023.3.post1
pywin32==308
PyYAML==6.0.2
pyzmq==26.2.0
redis==5.2.1
requests==2.32.3
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.37
sqlparse==0.4.4
stack-data==0.6.3
starlette==0.45.3
tornado==6.4.2
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2023.4
ujson==5.10.0
urllib3==2.3.0
uvicorn==0.34.0
watchfiles==1.0.4
wcwidth==0.2.13
websockets==14.2

# Added for advanced auth system
django-environ==0.11.2

# Better password hashing
argon2-cffi==23.1.0

# For production deployments
gunicorn==21.2.0
whitenoise==6.6.0